Move 'bool_op' predicates and manipulation into adiar::internal
