+++ src/adiar/internal/algorithms/prod2.h
@@ -4,10 +4,11 @@
 #include <variant>
 
 #include <adiar/access_mode.h>
-#include <adiar/bool_op.h>
 #include <adiar/memory_mode.h>
 #include <adiar/bdd/bdd.h>
+
 #include <adiar/internal/assert.h>
+#include <adiar/internal/bool_op.h>
 #include <adiar/internal/cnl.h>
 #include <adiar/internal/cut.h>
 #include <adiar/internal/dd.h>
